% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_movie.r
\name{get_movie}
\alias{get_movie}
\title{Get information for a specific movie}
\usage{
get_movie(title)
}
\arguments{
\item{title}{string}
}
\value{
list
}
\description{
Get information for a specific movie
}
\examples{
movieInfo <- get_movie("The Matrix")
}

