% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omdb_item_directors.R
\name{get_omdb_item_directors}
\alias{get_omdb_item_directors}
\title{Get OMDB Item Directors}
\usage{
get_omdb_item_directors(omdb_id)
}
\arguments{
\item{omdb_id}{String with the omdb_id for a movie/series.}
}
\value{
Return a vector with the directors of a movie/series.
}
\description{
This function returns a vector with the directors returned by the api (separated by commas).
}
\examples{
\dontrun{
get_omdb_item_directors(omdb_id = 'tt0120338')
}
}
\author{
Alberto Almuiña
}
