% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_omdb_item_ratings.R
\name{get_omdb_item_ratings}
\alias{get_omdb_item_ratings}
\title{Get OMDB Item Ratings}
\usage{
get_omdb_item_ratings(omdb_id, API_KEY = Sys.getenv("API_KEY"))
}
\arguments{
\item{omdb_id}{String with the omdb_id for a movie/series.}

\item{API_KEY}{OMBD Api Key. Default: Get the Api Key from system environment. Use Sys.setenv('API_KEY' = 'XXXXX'). More information in: http://www.omdbapi.com/apikey.aspx}
}
\value{
Return a tibble with the movie/series ratings.
}
\description{
This function searches Item Ratings.
}
\examples{
\dontrun{
get_omdb_item_ratings(odmb_id = 'tt0120338')
}
}
\author{
Alberto Almuiña
}
