% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_request_status.R
\name{get_request_status}
\alias{get_request_status}
\title{Get Request Status}
\usage{
get_request_status(res)
}
\arguments{
\item{res}{OMDB API Request}
}
\value{
Raise an error if the request is not correct.
}
\description{
This function determines if a OMDB API Request is correct or not.
}
\author{
Alberto Almuiña
}
