% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/item_to_database.R
\name{item_to_database}
\alias{item_to_database}
\title{Item to Database}
\usage{
item_to_database(con, item, dbtable, rownames = F, colnames = F,
  append = T)
}
\arguments{
\item{con}{Conecction to some database made with the RODBC package or some other.}

\item{item}{Tibble returned for some of the functions of the ROMDB package.}

\item{dbtable}{Name of the database table to insert the results.}

\item{rownames}{either logical or character. If logical, save the row names as the first column rownames in the table? If character, the column name under which to save the rownames. Default: FALSE}

\item{colnames}{logical: save column names as the first row of table? Default: FALSE}

\item{append}{logical. Should data be appended to an existing table? Default: TRUE}
}
\description{
This function stores an item into a database table.
}
\examples{
\dontrun{
item_to_database(con, tibble_df, 'M_SQL_TABLE')
}
}
\author{
Alberto Almuiña
}
