% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_movie.r
\name{search_movie}
\alias{search_movie}
\title{Search for a movie in the database}
\usage{
search_movie(search_string)
}
\arguments{
\item{search_string}{string}
}
\value{
tbl_df
}
\description{
Search for a movie in the database
}
\examples{
search_movie("The Matrix")
}

